+++ packages/zosjobs/__tests__/__unit__/GetJobs.unit.test.ts
@@ -332,6 +332,34 @@ describe("GetJobs tests", () => {
                 });
             GetJobs.getJobsCommon(pretendSession, {owner: "fakeOwner", prefix: "fakePrefix", maxJobs: 2, jobid: "fakeID"});
         });
+
+        it("should get a list of jobs from getJobsCommon with their execution data", async () => {
+            (ZosmfRestClient.getExpectJSON as any) = mockGetJobsJSONData([GetJobsData.SAMPLE_COMPLETE_JOB, GetJobsData.SAMPLE_ACTIVE_JOB]);
+
+            const jobCommon = await GetJobs.getJobsCommon(pretendSession, {execData: true});
+            expect(jobCommon).toMatchSnapshot();
+        });
+
+        it("should get a list of jobs from getJobsCommon with their execution data with maxJobs param", async () => {
+            (ZosmfRestClient.getExpectJSON as any) = mockGetJobsJSONData([GetJobsData.SAMPLE_COMPLETE_JOB, GetJobsData.SAMPLE_ACTIVE_JOB]);
+
+            const jobCommon = await GetJobs.getJobsCommon(pretendSession, {maxJobs: 2, execData: true});
+            expect(jobCommon).toMatchSnapshot();
+        });
+
+        it("should get a list of jobs from getJobsCommon with their execution data with owner and maxJobs params", async () => {
+            (ZosmfRestClient.getExpectJSON as any) = mockGetJobsJSONData([GetJobsData.SAMPLE_COMPLETE_JOB, GetJobsData.SAMPLE_ACTIVE_JOB]);
+
+            const jobCommon = await GetJobs.getJobsCommon(pretendSession, {owner: "someOwner", maxJobs: 2, execData: true});
+            expect(jobCommon).toMatchSnapshot();
+        });
+
+        it("should get a list of jobs from getJobsCommon with their execution data with multiple other params", async () => {
+            (ZosmfRestClient.getExpectJSON as any) = mockGetJobsJSONData([GetJobsData.SAMPLE_COMPLETE_JOB, GetJobsData.SAMPLE_ACTIVE_JOB]);
+
+            const jobCommon = await GetJobs.getJobsCommon(pretendSession, {owner: "someOwner", prefix: "fakePrefix", maxJobs: 2, execData: true});
+            expect(jobCommon).toMatchSnapshot();
+        });
     });
 
     describe("getJcl APIs", () => {